pyproject.toml
[project]
name = "prospect-presentation"
version = "0.1.0"
description = "Automated prospect presentation generator using RAG + local LLM"
requires-python = ">=3.10"

dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",

  "beautifulsoup4>=4.12.0",
  "requests>=2.32.0",

  "python-dotenv>=1.0.1",
  "pydantic>=2.8.0",

  "langchain>=0.1.0",
  "langchain-community>=0.0.20",
  "langchain-huggingface>=0.0.3",
  "langchain-text-splitters>=0.0.1",
  "chromadb>=0.4.0",
  "sentence-transformers>=2.2.0",
  "transformers>=4.30.0",
  "torch>=2.0.0",
  "llama-cpp-python==0.2.74",
  "accelerate>=0.27.0",
  
  "python-pptx",
  "loguru",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]